[Lexer] Lexer + Error handling

+++ lib/common/base.cc
@@ -55,7 +55,7 @@ auto utils::fits_in_b8(i64 num) -> i1 { return fits_in_i8(num) or fits_in_u8(num
 auto utils::fits_in_b16(i64 num) -> i1 { return fits_in_i16(num) or fits_in_u16(num); }
 auto utils::fits_in_b32(i64 num) -> i1 { return fits_in_i32(num) or fits_in_u32(num); }
 
-auto utils::number_width(u64 num, u32 base) -> u32 {
+auto utils::number_width(u64 num, u32 base) -> u8 {
     return num == 0 ? 1 : u32(std::log(num) / std::log(base) + 1);
 }
